Fractured block package
