+++ internal/producer/src/producer/producer_test.go
@@ -15,13 +15,13 @@ import (
 	"os"
 	"reflect"
 	"testing"
-	"time"
 
 	"github.com/SIGBlockchain/project_aurum/internal/client/src/client"
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accounts"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/block"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/blockchain"
+	"github.com/SIGBlockchain/project_aurum/internal/producer/src/hashing"
 	"github.com/SIGBlockchain/project_aurum/pkg/keys"
 )
 
@@ -91,7 +91,7 @@ func TestCheckConnectivity(t *testing.T) {
 func TestRunServer(t *testing.T) {
 	senderPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
 	recipientPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	recipientPublicKeyHash := block.HashSHA256(keys.EncodePublicKey(&recipientPrivateKey.PublicKey))
+	recipientPublicKeyHash := hashing.New(keys.EncodePublicKey(&recipientPrivateKey.PublicKey))
 	contract, _ := accounts.MakeContract(1, senderPrivateKey, recipientPublicKeyHash, 1000, 1)
 	contract.SignContract(senderPrivateKey)
 	serializedContract, err := contract.Serialize()
@@ -208,7 +208,7 @@ func TestByteChannel(t *testing.T) {
 	}
 	senderPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
 	recipientPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	recipientPublicKeyHash := block.HashSHA256(keys.EncodePublicKey(&recipientPrivateKey.PublicKey))
+	recipientPublicKeyHash := hashing.New(keys.EncodePublicKey(&recipientPrivateKey.PublicKey))
 	contract, _ := accounts.MakeContract(1, senderPrivateKey, recipientPublicKeyHash, 1000, 1)
 	contract.SignContract(senderPrivateKey)
 	serializedContract, _ := contract.Serialize()
@@ -355,7 +355,7 @@ func TestResponseToAccountInfoRequest(t *testing.T) {
 
 func TestData_Serialize(t *testing.T) {
 	senderPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	spkh := block.HashSHA256(keys.EncodePublicKey(&senderPrivateKey.PublicKey))
+	spkh := hashing.New(keys.EncodePublicKey(&senderPrivateKey.PublicKey))
 	initialContract, _ := accounts.MakeContract(1, nil, spkh, 1000, 0)
 	tests := []struct {
 		name string
@@ -409,7 +409,7 @@ func TestData_Serialize(t *testing.T) {
 
 func TestData_Deserialize(t *testing.T) {
 	senderPrivateKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-	spkh := block.HashSHA256(keys.EncodePublicKey(&senderPrivateKey.PublicKey))
+	spkh := hashing.New(keys.EncodePublicKey(&senderPrivateKey.PublicKey))
 	initialContract, _ := accounts.MakeContract(1, nil, spkh, 1000, 0)
 	// someData := &Data{
 	// 	Hdr: DataHeader{
@@ -449,80 +449,17 @@ func TestData_Deserialize(t *testing.T) {
 	}
 }
 
-func TestCreateBlock(t *testing.T) {
-	var datum []accounts.Contract
-	for i := 0; i < 12; i++ {
-		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := block.HashSHA256(keys.EncodePublicKey(&someKey.PublicKey))
-		someAirdropContract, _ := accounts.MakeContract(1, nil, someKeyPKHash, 1000, 0)
-		datum = append(datum, *someAirdropContract)
-	}
-	var serializedDatum [][]byte
-	for i := range datum {
-		serialData, _ := datum[i].Serialize()
-		serializedDatum = append(serializedDatum, serialData)
-	}
-	type args struct {
-		version      uint16
-		height       uint64
-		previousHash []byte
-		data         []accounts.Contract
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    block.Block
-		wantErr bool
-	}{
-		{
-			args: args{
-				version:      1,
-				height:       0,
-				previousHash: make([]byte, 32),
-				data:         datum,
-			},
-			wantErr: false,
-			want: block.Block{
-				Version:        1,
-				Height:         0,
-				Timestamp:      time.Now().UnixNano(),
-				PreviousHash:   make([]byte, 32),
-				MerkleRootHash: block.GetMerkleRootHash(serializedDatum),
-				Data:           serializedDatum,
-				DataLen:        12,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := CreateBlock(tt.args.version, tt.args.height, tt.args.previousHash, tt.args.data)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("CreateBlock() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got.Version, tt.want.Version) ||
-				!reflect.DeepEqual(got.Height, tt.want.Height) ||
-				!reflect.DeepEqual(got.PreviousHash, tt.want.PreviousHash) ||
-				!reflect.DeepEqual(got.MerkleRootHash, tt.want.MerkleRootHash) ||
-				!reflect.DeepEqual(got.DataLen, tt.want.DataLen) ||
-				!reflect.DeepEqual(got.Data, tt.want.Data) {
-				t.Errorf("CreateBlock() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func TestBringOnTheGenesis(t *testing.T) {
 	var pkhashes [][]byte
 	var datum []accounts.Contract
 	for i := 0; i < 100; i++ {
 		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := block.HashSHA256(keys.EncodePublicKey(&someKey.PublicKey))
+		someKeyPKHash := hashing.New(keys.EncodePublicKey(&someKey.PublicKey))
 		pkhashes = append(pkhashes, someKeyPKHash)
 		someAirdropContract, _ := accounts.MakeContract(1, nil, someKeyPKHash, 10, 0)
 		datum = append(datum, *someAirdropContract)
 	}
-	genny, _ := CreateBlock(1, 0, make([]byte, 32), datum)
+	genny, _ := block.New(1, 0, make([]byte, 32), datum)
 	type args struct {
 		genesisPublicKeyHashes [][]byte
 		initialAurumSupply     uint64
@@ -578,7 +515,7 @@ func TestAirdrop(t *testing.T) {
 	var pkhashes [][]byte
 	for i := 0; i < 100; i++ {
 		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := block.HashSHA256(keys.EncodePublicKey(&someKey.PublicKey))
+		someKeyPKHash := hashing.New(keys.EncodePublicKey(&someKey.PublicKey))
 		pkhashes = append(pkhashes, someKeyPKHash)
 	}
 	genny, _ := BringOnTheGenesis(pkhashes, 1000)
@@ -727,7 +664,7 @@ func TestRecoverBlockchainMetadata(t *testing.T) {
 	var pkhashes [][]byte
 	for i := 0; i < 100; i++ {
 		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := block.HashSHA256(keys.EncodePublicKey(&someKey.PublicKey))
+		someKeyPKHash := hashing.New(keys.EncodePublicKey(&someKey.PublicKey))
 		pkhashes = append(pkhashes, someKeyPKHash)
 	}
 	genny, _ := BringOnTheGenesis(pkhashes, 1000)
@@ -852,7 +789,7 @@ func TestRecoverBlockchainMetadata_TwoBlocks(t *testing.T) {
 	somePVKeys := make([]*ecdsa.PrivateKey, 3) // Grab 3 private keys for creating contracts
 	for i := 0; i < 100; i++ {
 		someKey, _ := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
-		someKeyPKHash := block.HashSHA256(keys.EncodePublicKey(&someKey.PublicKey))
+		someKeyPKHash := hashing.New(keys.EncodePublicKey(&someKey.PublicKey))
 		pkhashes = append(pkhashes, someKeyPKHash)
 		if i < 3 {
 			somePVKeys[i] = someKey
@@ -874,36 +811,36 @@ func TestRecoverBlockchainMetadata_TwoBlocks(t *testing.T) {
 	contracts := make([]accounts.Contract, 3)
 
 	// Contract 1
-	recipPKHash := block.HashSHA256(keys.EncodePublicKey(&(somePVKeys[1].PublicKey)))
+	recipPKHash := hashing.New(keys.EncodePublicKey(&(somePVKeys[1].PublicKey)))
 	contract1, _ := accounts.MakeContract(1, somePVKeys[0], recipPKHash, 5, 1) // pkh1 to pkh2
 	contract1.SignContract(somePVKeys[0])
 	err = accounts.ValidateContract(contract1)
 	if err != nil {
 		t.Error(err.Error())
 	}
-	senderPKHash := block.HashSHA256(keys.EncodePublicKey(&(somePVKeys[0].PublicKey)))
+	senderPKHash := hashing.New(keys.EncodePublicKey(&(somePVKeys[0].PublicKey)))
 	accounts.ExchangeBetweenAccountsUpdateAccountBalanceTable(acctsDB, senderPKHash, recipPKHash, 5) // update accts table for further contracts
 
 	// Contract 2
-	recipPKHash = block.HashSHA256(keys.EncodePublicKey(&(somePVKeys[2].PublicKey)))
+	recipPKHash = hashing.New(keys.EncodePublicKey(&(somePVKeys[2].PublicKey)))
 	contract2, _ := accounts.MakeContract(1, somePVKeys[1], recipPKHash, 7, 2) // pkh2 to pkh3
 	contract2.SignContract(somePVKeys[1])
 	err = accounts.ValidateContract(contract2)
 	if err != nil {
 		t.Error(err.Error())
 	}
-	senderPKHash = block.HashSHA256(keys.EncodePublicKey(&somePVKeys[1].PublicKey))
+	senderPKHash = hashing.New(keys.EncodePublicKey(&somePVKeys[1].PublicKey))
 	accounts.ExchangeBetweenAccountsUpdateAccountBalanceTable(acctsDB, senderPKHash, recipPKHash, 7) // update accts table for further contracts
 
 	// Contract 3
-	recipPKHash = block.HashSHA256(keys.EncodePublicKey(&(somePVKeys[1].PublicKey)))
+	recipPKHash = hashing.New(keys.EncodePublicKey(&(somePVKeys[1].PublicKey)))
 	contract3, _ := accounts.MakeContract(1, somePVKeys[2], recipPKHash, 5, 2) // pkh3 to pkh2
 	contract3.SignContract(somePVKeys[2])
 	err = accounts.ValidateContract(contract3)
 	if err != nil {
 		t.Error(err.Error())
 	}
-	senderPKHash = block.HashSHA256(keys.EncodePublicKey(&somePVKeys[2].PublicKey))
+	senderPKHash = hashing.New(keys.EncodePublicKey(&somePVKeys[2].PublicKey))
 	accounts.ExchangeBetweenAccountsUpdateAccountBalanceTable(acctsDB, senderPKHash, recipPKHash, 5) // update accts table
 	acctsDB.Close()
 
@@ -911,7 +848,7 @@ func TestRecoverBlockchainMetadata_TwoBlocks(t *testing.T) {
 	contracts[1] = *contract2
 	contracts[2] = *contract3
 
-	firstBlock, err := CreateBlock(1, 1, block.HashBlock(genesisBlk), contracts)
+	firstBlock, err := block.New(1, 1, block.HashBlock(genesisBlk), contracts)
 	if err != nil {
 		t.Errorf("failed to create first block")
 	}
@@ -962,7 +899,7 @@ func TestRecoverBlockchainMetadata_TwoBlocks(t *testing.T) {
 			}()
 
 			for i, key := range somePVKeys {
-				someKeyPKhsh := block.HashSHA256(keys.EncodePublicKey(&key.PublicKey))
+				someKeyPKhsh := hashing.New(keys.EncodePublicKey(&key.PublicKey))
 				var balance uint64
 				var nonce uint64
 				queryStr := fmt.Sprintf("SELECT balance, nonce FROM account_balances WHERE public_key_hash=\"%s\"", hex.EncodeToString(someKeyPKhsh))
